For debug

+++ repair.hpp
@@ -65,7 +65,7 @@ typedef struct Dictionary {
   size_t txt_len;
   uint num_rules;
   RULE *rule;
-  uint seq_len;
+  size_t seq_len;
   CODE *comp_seq;
   uint buff_size;
 } DICT;
